Add support for driver safe loading
Support for safe driver loading is implemented as
a part of the upgrade flow.
When UpgradeStateManager detects a node that is waiting
for a safe driver load, it will unconditionally transfer
it to the UpgradeStateUpgradeRequired state and wait for Cordon
and Drain operations to complete according to the upgrade policy.
When the Pod is eventually in the UpgradeStatePodRestartRequired state,
the UpgradeStateManager will unblock the driver loading
(by removing the safe driver load annotation)
instead of restarting the Pod.
The default implementation of the SafeDriverLoadManager interface
assumes that the driver's safe load mechanism is implemented as
a two-step procedure.
As a first step, the driver pod should load the init container,
which will set "safe driver load annotation"
(defined in UpgradeWaitForSafeDriverLoadAnnotationKeyFmt)
on the node object, then the container blocks until another
entity removes the annotation from the node object.
When the init container completes successfully
(when the annotation was removed from the Node object),
the driver Pod will proceed to the second step and do the driver loading.
After that, the UpgradeStateManager will wait for the driver
 to become ready and then Uncordon the node if required.